refactoring database connection to config folder. Separating Scraper and Rest endpoint files
